(PC-25501) refactor(offer): move the code specific to the v2 offer into a specific folder
