+++ src/features/offerv2/components/OfferArtists/OfferArtists.tsx
