+++ src/features/offerv2/components/OfferPrice/OfferPrice.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 import React from 'react'
 
-import { OfferPrice } from 'features/offer/components/OfferPrice/OfferPrice'
+import { OfferPrice } from 'features/offerv2/components/OfferPrice/OfferPrice'
 
 const meta: ComponentMeta<typeof OfferPrice> = {
   title: 'features/offer/OfferPrice',
